Free memory that wasn't being freed.
ximage.data wasn't being freed and part of the problem was that it was being increased by linelen, so the offset would be wrong, corrected this.  Also in notify() apptext was being set to "xace" without any storage being allocated, corrected this.  Finally in notify() appname and iconname were not being freed.