Add proper checksum when saving files
I have only added the checksum when saving files, as with digital storage there should be little need to check the checksum when loading files, plus this allows old files to be easily updated by just loading them and saving again.  That said it may be useful later, via a command switch.